Move ServerMessage, Credentials, AccesToken, and SwapEvent models to model folder.
