+++ gridt/src/app/core/api.service.ts
@@ -4,11 +4,8 @@ import { Observable, BehaviorSubject, throwError, merge, partition } from "rxjs"
 import { map, tap, pluck, catchError, flatMap, distinctUntilChanged, take } from "rxjs/operators";
 import { Movement } from "./models/movement.model";
 import { User } from './models/user.model';
+import { ServerMessage } from './models/servermessage.model';
 import { AuthService } from './auth.service';
-
-export interface ServerMessage {
-  message: string;
-}
 
 @Injectable({
   providedIn: "root"
