+++ gridt/src/app/core/models/credentials.model.ts
@@ -1,0 +1,8 @@
+export interface Credentials {
+  username: string;
+  password: string;
+}
+
+export interface AccessToken {
+  access_token: string;
+}
