+++ gridt/src/app/core/models/servermessage.model.ts
@@ -1,0 +1,3 @@
+export interface ServerMessage {
+  message: string;
+}
