+++ gridt/src/app/core/models/swapevent.model.ts
@@ -1,0 +1,8 @@
+import { Movement } from './movement.model';
+import { User } from './user.model';
+
+export interface SwapEvent {
+  date: Date;
+  movement: Movement;
+  user: User;
+}
